A slightly more interesting station spawn system

+++ src/randoms.rs
@@ -2,15 +2,15 @@ use std::sync::mpsc::Receiver;
 use rand::{Rng, thread_rng};
 
 pub trait Random {
-    fn gen(&self) -> f32;
+    fn gen(&self) -> f64;
 }
 
 pub struct TestRandom {
-    pub r: Receiver<f32>,
+    pub r: Receiver<f64>,
 }
 
 impl Random for TestRandom {
-    fn gen(&self) -> f32 {
+    fn gen(&self) -> f64 {
         self.r.recv().unwrap()
     }
 }
@@ -18,7 +18,7 @@ impl Random for TestRandom {
 pub struct RealRandom;
 
 impl Random for RealRandom {
-    fn gen(&self) -> f32 {
+    fn gen(&self) -> f64 {
         thread_rng().gen()
     }
 }
